Moved the test folders
